backend/requirements.txt
google-adk>=1.20.0
ag-ui-adk>=0.3.5
fastapi>=0.115.0,<0.119.0
uvicorn>=0.34.0
python-dotenv>=1.0.0